Q and A rule evaluation

+++ src/main/java/com/facilio/qa/context/AnswerHandler.java
@@ -5,7 +5,6 @@ import com.facilio.qa.rules.pojo.ScoringRuleCondition;
 import com.facilio.time.DateRange;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-import org.apache.commons.collections4.CollectionUtils;
 
 import java.io.Serializable;
 import java.util.List;
@@ -33,7 +32,7 @@ public abstract class AnswerHandler<A extends ClientAnswerContext> implements Se
 
     }
 
-    public int computeFullScore (List<ScoringRuleCondition> conditions) {
+    public double computeFullScore (List<ScoringRuleCondition> conditions) {
         return ScoringRule.computeMaxScore(conditions);
     }
